requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
python-dotenv==1.0.1
apscheduler==3.10.4
claude-agent-sdk==0.1.3
anyio==4.6.0
